Renamed routes to comply with REST naming conventions
